requirements.txt
# requirements.txt
selenium
pytest
PyYAML
webdriver-manager